pipeline: neo: Rename INPUT[N] streams as IMAGE[N] stream
That change is non functional.

The camera V4L2 streams related to image data are currently named as
INPUT0 and INPUT1. Reason is that so far, those streams have been related
to HDR use case whereby they correspond to long and short captures and
are passed to the ISP INPUT0 and INPUT1 video nodes.

However, the 2 image streams for the camera are not necessarily passed to
INPUT0 and INPUT1 video nodes. For RGBIr context switch, both streams have
to be input sequentially to the INPUT0 device of the ISP. For clarity,
rename those streams as IMAGE0 and IMAGE1 which is also more consistent
with the internal pad naming in the sensor subdevices.
While at it, rename the streams enum from capital to camel case to better
match the style of the code base.

Also update all symbols in the pipeline to use this new convention.
Explicit references to one or the other of the ISP inputs are not renamed
as they do not represent a camera image stream.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>