pipeline: neo: Move raw stream only indication to NxpNeoFrames class
The `raw stream only` information means that there is no stream to be
decoded by the ISP so that both ISP and IPA can be bypassed.
This argument is currently passed at each NxpNeoFrames::Info instance
creation which is allocated for each frame to be captured. That is not
necessary as that information is global to the camera session and does
not change from frame to frame. Thus, move this argument to the
NxpNeoFrames instance instead, which is created once when streaming
session starts.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>