pipeline: neo: Rename BufferType enum values
This change renames BufferType enum values from Type[xyz] to
BufferType[xyz] instead of Type[xyz] so that it becomes explicit from
their name what they represent.

That is in prevision of adding more enums in the mojom interface.
Also while at it rename input0/iput1 buffer types as image0/image1 for
consistency with the implementation.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>