ipa: neo: Add gamma control algorithm
This algorithm configures the GCM block of the ISP.
This block performs a gamma correction on input pixels.

It is using the user color space to configure the parameters
of following operations based on the standards definition:
- the transfer function for the gamma correction
- the Y'CbCr encoding for the R'G'B' to Y'U'V' conversion

There is also the option to configure the gamma value of the transfer
function using the calibration file.

The algorithm is providing the gamma value to the user metadata.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>