pipeline: neo: Use of StreamType as an enum
The StreamType enum represents a camera stream, either an image or
embedded data, typically sent on a separate MIPI-CSI2 Virtual Channel.

This is a non-functional change that updates the implementation to
use StreamType as an enum instead of its underlying integer type.
That is both for readability and enum type values checking.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>