ipa: neo: drc: Set dynamic DRC control for ox03c10 sensor
The DRC control for the ox03c10 is set as dynamic control.
For this control, output results is based equally on histogram
equalization and histogram stretching (gdrc-alpha=128).
The strength of the dynamic range compression with histogram
stretching is set to 180 (gdrc-gamma)
The default compression curve is using a fixed gamma set to 180.

Signed-off-by: Martin Slanina <martin.slanina@nxp.com>