ipa: neo: lsc: Remove vignetting lut for ct=0K
For the Lens Shading Correction, tuning of the LUTs
is provided for 3 illuminant modes:
- Halo: <2800K,
- Fluorescent (indoor/LED): [3600K-4200K]
- D65 (daylight): >5000K
Hence LUTs should be defined for each of those color temperature point:
2800K, 3600K, 4200K and 5000K
The LUT for 0K which is similar to the 2800K can be removed.
Indeed no interpolation is performed for color temperature
below the lowest point and above the highest point.

This patch helps to reduce tuning file parsing time.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>