ipa: neo: rgbir: Correct detection of IR output format
This patch fixes the detection of the IR output format
8/16 bits.
The existing check of format was applying to all the streams.
However if among the streams, the frame stream is configured to
the monochrome format Y8 or Y16, this results in wrong detection.

This patch is now checking the format for the IR stream only.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>