pipeline: neo: Add ISP and IPA usage in raw-only mode
Current implementation of the raw-only mode (raw stream enabled with no
stream for ISP decoded image) is that ISP decoding and IPA processing
are bypassed from the pipeline. The consequence is that there is no
exposure control of the raw image and that there is also no metadata
produced by the IPA to bind to that buffer.

This change amends the pipeline handler operation in raw-only mode to
still have the raw images decoded by the ISP and have the IPA running
normally for ISP and sensor control.
For that purpose, even though libcamera::Request contains no buffer from
the application for a decoded image output, the ISP frame output buffer
video node is configured and enabled and internal buffers allocated for
its operation.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>