pipeline: imx8-isi: Delay ISI routes config to acquire() time
Fixes behavior when calling 'cam -l' during a live stream from
a camera in another process.
Issue is that multiple process should be able to list (match
procedure) the camera supported. But only the unique process
that lock the media devices in order to be able to configure
then start the pipeline should setup the routes, graphs etc.
Thus, the setRouting() is to be moved to a
libcamera::PipelineHandlerISI::acquireDevice() implementation
to override the default Pipeline::acquireDevice() function.

Tested-by: Antoine Bouyer <antoine.bouyer@nxp.com>
Fixes: 92df79112fb2 ("pipeline: imx8-isi: Add multicamera support")

Signed-off-by: Andrei Gansari <andrei.gansari@nxp.com>