ipa: neo: Fix infrared stream only operation
Current implementation segfaults when running a pipeline for a RGBIr
sensor with a single stream for IR without RGB/YUV stream.
Reason is that IPA accesses IPAStream entry from a vector of streams
configuration assuming that both streams are populated.

This change stores the IPAStream configuration locally in the IPA in a
map of IPAStreamType, as passed over the IPA interface. That way, the
presence of the IR stream configuration entry can be tested before its
dereference.

Fixes: 2e7fde0ae2e7 ("ipa: neo: rgbir: Correct detection of IR output format")
Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>