libamera: media_pipeline: Move entities list to public
Exposes internal MediaEntity::Entity list to help extracting more
information regarding linked entities.

For example, when the pad index of the last device in the list need to
be retrieved from the media pipeline user.

Exposes as const to prevent any corruption from user. Then it is still
protected so as when the list was private.

Signed-off-by: Andrei Gansari <andrei.gansari@nxp.com>
Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>