pipeline: neo: Fixes IPA buffersId keys usage
Pipeline fills in bufferIds maps passed to the IPA in the
computeParams() and processStats() IPC calls.
The keys of those maps are expected to be IPABufferTypeStats (from IPA
mojom interface) whereas the pipeline populates the maps using
BufferType enum that is local to the pipeline handler.
Current implementation works because those enums match at the moment
but they may diverge at some point.

This is a non-functional change, updating the pipeline handler to use
the IPABufferTypeStats enum as needed for bufferIds initialization.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>