pipeline: neo: Move sensor configuration during ISI image receipt
Sensor controls configuration is currently triggered on the ISP frame
sync event, meaning that the image decoding job has started.
Intent is to synchronize the sensor configuration to be always done at
the same position during the frame period. The limitation of this
trigger is in multi-camera operation: there is a single hardware ISP so
images decoding has to be serialized. This induces a delay for the
different cameras to be individually reconfigured.

To overcome this limitation, the camera reconfiguration is moved to the
ISI channel receipt trigger, which is not sensible to the ISP
access contention.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>