pipeline: neo: Fix processStats() control
The common sequence number used for NxpNeoFrames::Info, IPA frame
number and delayedControls comes from the libcamera::Request sequence
number. That sequence number is most of the time identical to the v4l2
sequence number from front-end or ISP, but may differ e.g. in case of
frame loss.

This change fixes the sequence number used to retrieve the control list
in use in the sensor, that is passed to IPA in the processStats() IPC
call. Use the correct sequence number from the libcamera::Request in
there as well for robustness.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>