ipa: neo: Set the camera mode from init()
At initialization, the camera mode needs to be set with the
current sensor mode, otherwise the line duration is computed
with non-initialized parameters.
This patch defines a function updateSensorConfig() to set the
camera mode and to update the IPA context which needs to be
performed from init() and from configure().

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>