pipeline: neo: Instantiate DelayedControls object per context
DelayedControls object tracks the sensor controls configuration updates
and accounts for individual propagation delay so that the changes for a
group of controls are applied simultaneously in the sensor.
For that purpose, the controls with a long propagation delay are sent in
advance compared to the controls with a shorter delay.

Therefore, the actual controls values sent to the sensor for a given
frame is a mix of controls picked from the past updates according to
their individual delay. This does not play well with RGBIr context
switch mode of operation where controls updates are produced
alternately for RGB and IR contexts. Updates have to be stored in
separate DelayedControl instances, so that the delayed-adjusted updates
sent to the sensor are produced only with control values for the
targeted context.

For that purpose, that change instantiates the DelayedControl object for
each context.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>