ipa: neo: pipe_conf: Set a LPALIGN default value dependency on mode
In HDR mode, PIPECONF.LPALIGN0/1 should be set to zero to keep the
native sensor pixel format to feed the HDR merge block.
This change updates the defaut configuration of the PIPECONF (i.e.
when not explicitly defined in the calibration file) to take into
account the pipeline mode. If the mode is HDR merge, LPALIGN0/1 default
values are set to zero, as known at configure() time.
In-source documentation is updated accordingly.
While at it, some minor edits have been applied in documentation for
improvement.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>