ipa: neo: Update camera mode with the sensor stream mode
The sensor stream mode types (standard, hdr, rgbIr, dual context) are
defined from the camera_helper interface.
The stream mode of the sensor is added part of the camera mode structure.
This is used by the cameraHelper to decide implementation according
to the stream mode.

This patch also adds the sensor bit depth and sensor size part of the
camera mode structure to compute the sensor timings.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>