ipa: neo: Use multi-capture controls in HDR mode for os08a20 sensor
In HDR mode, specific usage of the V4L2 multi-capture controls is implemented
in the cameraHelper of the os08a20.
This involves specific implementation for:
- controlListSetAGC() with the distribution of the exposure and gain
  for the short and long captures.
- sensorControlsToMetaData() to get sensor setting from the multi-capture controls
- controlInfoMapGetExposureRange() and controlInfoMapGetAnalogGainRange()
  to get min/max values from hard coded setting.

For this purpose, the controlListSetAGC() and sensorControlsToMetaData()
are relying on the V4L2_CID_EXPOSURE_MULTI and V4L2_CID_AGAIN_MULTI
controls which allows to provide exposure and gain for both captures.

The distribution function uses a ratio of 16 between the long and short
total exposure.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>