ipa: neo: blc: Update algorithm for HDR merge support
During HDR merge operation, native pixel format has to be used in the
ISP pipeline, from input0/input1 up to the HDR block.
As a consequence, for HDR merge operation, PIPECONF.LPALIGN0/1=0 has
to be configured to avoid rescaling and HDR Decomp use as bypass or
with fractional gain to counteract on the unwanted PIPECONF rescaling.

In non HDR-merge mode of operation, assumption is that after HDR Decomp
blocks, internal pixel format is either 20-bit (input0) or 16-bit
(input1), applicable to OBWB0/1 inputs.
Conversely, in HDR-merge mode of operation, assumption is that
native pixel format will be used up to to HDR Merge block, applicable to
OBWB0/1 inputs. Then 20-bit output format is expected from HDR merge
block output onward, applicable to OBWB2.

This change adds the logic to apply the appropriate ISP pipeline gain
to the actual sensor BLC offset, depending on the mode of operation.
Documentation is complemented to describe the pipeline configuration
for the different modes.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>