ipa: neo: hdr merge: Condition algorithm usage to HDR merge mode
In current implementation, HDR merge alorithm is enabled and the
relevant associated ISP block configured as long as it is defined in the
calibration file. If the pipeline is not running in HDR merge mode, but
the ISP block is configured, this induces kernel driver errors as the
input1 path is not active.

This change conditions the activation of the HDR merge block to the fact
that the pipeline handler is running in HDR merge mode. Purpose is to be
able to share the same calibration file for both HDR and non-HDR merge
modes.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>