ipa: neo: awb: Make awb default obwb selection mode-dependent
The current implementation defaults to the usage of obwb2 block selection
when not specified in the calibration file. However, for HDR-merge, the
default mode should be obwb0/1 instead.

This change updates the default obwb blocks selected in case of HDR
merge mode of operation to be obwb0/1 instead of obwb2.
The obwb output format (obpp) is also dynamically defined at configure()
time as its format varies depending on the mode.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>