libcamera: camera_sensor: Fix imageStream() for CameraSensorLegacy class
The CameraSensor::imageStream() has a default implementation in the base
class that returns (0, 0) as (pad, stream), assuming that image pad is
located on the pad index 0.
This assumption is correct most of the time, but not in some other
cases, for instance when an external ISP entity acts as the sensor.
Such entity would typically have sink pad(s) connected to the actual
sensor and source pad(s) connected to the downstream graph. Such source
pad indexes would likely be different from zero.

CameraSensorLegacy subclass correctly handles this case in its
functions, using the pad_ variable discovered at init() time to access
the source pad index.
Exception is imageStream() that is not overriden in the
CameraSensorLegacy definition so keeps the default implementation of
the parent class, hardcoding the returned source pad index to zero.

This change declares CameraSensor::imageStream() as a pure virtual to
let the subclasses provide an implementation. Implementation for
CameraSensorLegacy is added based on pad_ variable usage.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>