pipeline: neo: Add smart camera support
This change adds basic support for smart cameras in the neo pipeline,
exposing a native camera stream to the application, in one of the output
format permitted by the ISI CSC capabilities. Camera stream crop,
scaling and duplication is not implemented.

The main specificities introduced by the support of smart cameras are:
- Enumeration: criteria for sensor compatibiliy consists in the output
  format to be compatible with one of the processed input format of the
  ISI.
- StreamConfiguration handling: a single stream is supported for now
  with one of the native sensor resolution as no duplication and no
  crop/rescaling is implemented. This single stream is currently mapped
  to the streamRaw for simplicity.
- Pipeline handler stages: the output buffers from the ISI pipes are
  provided by the application as part of the libcamera::Request. Thus,
  there is no internal buffer allocation required. The completion of the
  libcamera::Request happens earlier than with raw camera since there is
  no IPA and no ISP processing involved.

From implementation standpoint, multiple functions have been
instantiated into `Raw` and `Yuv` versions to avoid addition of large
if-else specialization blocks in the existing functions.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>