pipeline: neo: Fix neo output pixel format conversion
A PixelFormat can be converted to one or multiple V4L2PixelFormat. The
actual V4L2PixelFormat to use depends on the underlying V4L2VideoDevice
capabilities.

The NxpNeoCameraData::configureRaw() arbitrarily convers the PixelFormat
from the StreamConfiguration into the first V4L2PixelFormat value
reported by V4L2PixelFormat::fromPixelFormat() conversion functions.
Though functional, that is not the correct way of converting that should
enquire the video device instead about which corresponding
V4L2PixelFormat it supports via V4L2VideoDevice::toV4L2PixelFormat().

This implements the above adaptation for both frame and IR output of the
ISP.
While at it, the access to the colorSpace variable of the StreamConfig
object is explicited as being a std::optional<> for readability.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>