ipa: neo: ox03c10: Calculate VTS based on height and vblank
The VTS value is calculated based on the sensor height and vblank.
This helps to remove the hard coded value for VTS and maximum
exposure in double lines.
For the 1920*1280 resolution, the calculation of the maximum exposure
in double lines gives 641 lines versus the hard coded value of 631 lines
(675-31-12-1).

This patch allows to have the VTS and maximum exposure in double lines
to be computed depending on the sensor resolution.

The constants min/max exposure in double lines are renamed explicitly
with "DoubleLines" as suffix in the constant name.

This patch also:
- removes the unused convertRows2ExposureTime()
- rename explicitly the constants min/max exposure in double lines
with "DoubleLines" as suffix in the constant name.
- rename local variables with "xxxdoubleRows" when its unit is double rows
- rename convertExposureTime2Rows() into convertExposureTime2DoubleRows()

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>