ipa: neo: Add input1 bpp in configuration context
Multiple IPA have a dependency on the sensor input bit per pixel value
for input0 and sometime input1. It could happen that a sensor use a
different bpp format for input0 and input1 image for HDR operation.
As of now input1 bpp is not reported by libcamera core. That is
currently not an issue as most HDR sensor use the same format for
both images.

Still, algorithms requiring access to both formats typically have a
local todo as they have to assume same pixel format for both inputs.
They also do not consistently access the input0 bpp value, either
caching it from IPACameraSensorInfo at configure() time, or accessing
the input0 value from the context IPASessionConfiguration struct.

This changes update the single bpp format in IPASessionConfiguration to
an array of bpps covering both inpu0 and input1. A single \todo is now
present to cover the duplicated input1 bpp value when this array is
initialized at configure() time.
The IPA algorithms are updated to use this bpps array from the context.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>