ipa: neo: hdr_decomp: Update HDR-merge 10-bit mode
In HDR merge mode of operation, AWB function is typically operated on
OBWB0 and OBWB1 ISP blocks. At that stage, the pixel format in the ISP
pipeline is the sensor native format. However, OBWB function requires
saturation to be configured for the appropriate format, configurable for
12-bit formats and higher.

Thus, for HDR-merge the HDR Decomp ISP block has to rescale the native
format on input0 and/or input1 from 10-bit to 12-bit when needed.

Also the rescaling logic is moved from prepare() function to configure()
where it should have been initially.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>