ipa: neo: blc: Update HDR-merge 10-bit mode
In HDR merge mode of operation, AWB function is typically operated on
OBWB0 and OBWB1 ISP blocks. At that stage, the pixel format in the ISP
pipeline is the sensor native format. However, OBWB function requires
saturation to be configured for the appropriate format, configurable for
12-bit formats and higher.

This change takes into account that for HDR merge with 10-bit input on
either and/or input1, there is a gain applied in HDR Decomp to rescale
format to 12-bit, impacting the offset values to be configured in OBWB1
and OBWB2 for the BLC correction.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>