pipeline: neo: Add controls update configuration
The event used to trigger the camera controls update has been moved from
the ISP sync event to the ISI frame available event, in order to provide
a more determistic timing during the frame period.

This change has created an adverse effect on mx95mbcam (metadata
corruption, image corruption) as sensor seems to struggle when updated
with that new event.
As a workaround, add an optional configuration in the pipeline
configuration file to still use the ISP sync event to trigger the
controls update. That is configurable per camera.

For that purpose, this change reintroduces the ISP frameStart()
callback. In order to detect the active context being operated, the
processingRequests_ queue is moved from std::queue to a std::deque so
that it can be iterated to find the current active context.
The target context is identified as being the first one from the
requests queue having its stats buffer available.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>