ipa: neo: os08a20: Enable dynamic DRC control
Enable the dynamic global DRC control for the os08a20 sensor
with following parameters:
- output results is based equally on histogram
equalization and histogram stretching (gdrc-alpha=128).
- strength of the dynamic range compression with histogram
stretching is set to 180 (gdrc-gamma)
- default compression curve is using a fixed gamma set to 180.

This patch also adds a constraint to the AGC with a target value
not to exceed as an upper bound. This helps to regulate the number
of nearly saturated pixels (top 2% of pixels) and lowers the
exposure in order to stop pixels saturating.
This improves very bright areas (such as bright outside window
captured from an indoor scene) to not saturate. This results in
texture and colours to become visible.
For information, the constraint lower bound was already applied as
default constraint.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>