pipeline: neo: Improve encapsulation of NxpNeoFrames classes
The NxpNeoFrames class manages the buffers and metadata necessary to
serve each libcamera::Request. It has 2 companion classes: Info that
embeds all data associated to a Request, and InfoContext that
embeds data associated to only one of the possibly multiple contexts
for a Request.
Helpers are provided for the pipeline handler to use those classes.
They are currently all associated to the top level NxpNeoFrames class.

This non functional change aims at improving the encapsulation of the
classes, binding the helpers directly to the underlying class
targetted by a given helper.
For that purpose, NxpNeoFrames::InfoContext declares and defines:
- completeBuffer()
- isBufferPending()
- isContextComplete()
- buffer()
And NxpNeoFrames::Info declares and defines:
- isFrameComplete()

Those functions are relocated accordingly in the source file to match
their declaration order.

Pipeline handler implementation is updated where those helpers are
used to reference the new class hosting the helper.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>