pipeline: neo: Move raw/yuv specialization functions to private
With the introduction of smart sensors in the neo pipeline, some of the
functions were specialized into yuv and raw variants. Existing functions
have been renamed from function() to functionRaw() and initially kept in
place along with the newly introduced functionYuv() to ease the review.

In the case where the initial function() was public, it is transformed as
a simple wrapper to call either of the specialized functions.
Thus, the specialized functions can be moved to private, and their
definitions relocated accordingly, as this change is doing.

For consistency with the other specialized functions,
generateRawConfiguration() and generateYuvConfiguration() are also
renamed to generateConfigurationRaw(), respectively
generateConfigurationYuv().

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>