ipa: neo: hdr: Add the ratioL2S parameter as user configuration
The ratio between the long and the short captures can be used by multiple
algorithms for computing ISP parameters, such as:
- the agc algorithm to adapt the STAT configuration
- the hdr algorithm to compute the HDR configuration

Hence the ratioL2S is created as a HDR calibration parameter.
It can be configured in the yaml file using the node "ratio-long2short".
The default value is set to 16.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>