ipa: neo: drc: Add possibility to restrict DRC in HDR merge mode only
This patch adds the possibility to restrict the DRC operation to run
only if the pipeline is in HDR merge mode.
Such restriction can be set from the calibration file
with the "restrict-mode" key:
The valid values are: { "hdr-merge", "none" }

If the restrict mode is set to "hdr-merge" and if the pipeline
is not in HDR merge mode, the DRC operation is disabled
(the DRC mode is forced to gbl-mode=0, Copy mode).

This patch moves the initialization of the LUT from configure()
to adapt in case the pipeline mode is changing from one session
to another one.

The LUT vector is changed as array type since the size is constant.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>