ipa: neo: os08a20: Remove upper bound constraint for the AGC
The AGC upper bound constraint was necessary with the DRC operation
in standard pipeline mode.
However since the dynamic global DRC is enabled for os08a20 only
in hdr merge mode, the upper bound contraint is not needed anymore.

Note that the AGC lower bound constraint is already defined as default
from the libipa.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>