ipa: neo: drc: Adapt initialization of the GDRC context
Following variables from gblDrcContext_ are reinitialized
from configureGblDrcContext() instead of initializeGblDrcContext():
- minValue/maxValue
- minBin/maxBin

Following variables from gblDrcContext_ are changed as constexpr:
- minPixelCount/maxPixelCount
- histEqThreshold
- histEqSaturation

Initialization of following variables from gblDrcContext_ are removed
from initializeGblDrcContext() since there are initialized at init()
from the calibration file:
- globalDrcAlpha (default value is 128)
- gammaFixed (default value is 140)
Also their const default values are changed to be aligned with the ones set
previously with initializeGblDrcContext().

With this change, the function initializeGblDrcContext() is not necessary
anymore and is removed.

Also the LUT vector passed as an argument to controlDynamicMode() and
lutSecondRun() is removed and the class member gblLut_ is used directly instead.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>